requirements.txt
python-telegram-bot==21.6
pillow==10.4.0
requests==2.32.3
psycopg2-binary==2.9.9
